id are int
git-svn-id: http://sbfsvn01.cl.gemini.edu/software/yorick-svipc/trunk@28504 ee6ba543-07f8-4773-b02c-01d50e9d00ba

+++ demo/pong.py
@@ -7,7 +7,7 @@ import time
 
 
 # retreive ids
-my_shmid,my_semid,my_msqid = [long(a) for a in sys.argv[1:]]
+my_shmid,my_semid,my_msqid = [int(a) for a in sys.argv[1:]]
 
 # wait one sec
